altaração do id de Integer para Long

+++ src/modelo/dao/CidadeDao.java
@@ -6,8 +6,6 @@
 package modelo.dao;
 
 import entidades.Cidade;
-import entidades.Cidade;
-import java.io.Serializable;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
@@ -18,6 +16,7 @@ import modelo.dao.uteis.ConexaoDB;
  * @author wender
  */
 public class CidadeDao {
+
     public Cidade novo() {
         return new Cidade();
     }
@@ -38,17 +37,17 @@ public class CidadeDao {
         em.getTransaction().begin();
         if (cidade.getId() != null) {
             em.remove(em.find(cidade.getClass(), cidade.getId()));
-        }        
+        }
         em.getTransaction().commit();
         em.close();
     }
 
-    public Cidade pesquisar(Integer id) {
-         EntityManager em = ConexaoDB.getEntityManager();
-         Cidade cidade = (Cidade)em.find(Cidade.class, id);
-         return cidade;
+    public Cidade pesquisar(Long id) {
+        EntityManager em = ConexaoDB.getEntityManager();
+        Cidade cidade = (Cidade) em.find(Cidade.class, id);
+        return cidade;
     }
-    
+
     public List<Cidade> pesquisar(String nome) {
         EntityManager em = ConexaoDB.getEntityManager();
         Query consulta = em.createNamedQuery("Cidade.findByNome");
@@ -56,6 +55,4 @@ public class CidadeDao {
         return consulta.getResultList();
     }
 
-    
-    
 }
